Extracted minifying operations to separate file
