+++ scripts/constants.js
@@ -1,0 +1,3 @@
+module.exports.BUILD_DIR = './build';
+module.exports.SRC_DIR = './src';
+module.exports.DATA_SRC_DIR = `${module.exports.SRC_DIR}/data`;
